style: removed code duplication in course querying
